// src/core/adapters/drizzle/schema.ts
import { relations, sql } from "drizzle-orm";
import { integer, sqliteTable, text } from "drizzle-orm/sqlite-core";
import { ulid } from "ulid";

export const primaryKeyAutoGenerated = () =>
  text("id")
    .primaryKey()
    .$defaultFn(() => ulid());

export const products = sqliteTable("product", {
  id: primaryKeyAutoGenerated(),
  name: text("name").notNull(),
  imageUrl: text("imageUrl").notNull(),
  price: integer("price").notNull(),
  categoryId: text("category_id").notNull(),
  createdAt: text("createdAt")
    .notNull()
    .default(sql`CURRENT_TIMESTAMP`),
});

export const productsRelations = relations(products, ({ one }) => ({
  category: one(category, {
    fields: [products.categoryId],
    references: [category.id],
  }),
}));

export const category = sqliteTable("category", {
  id: primaryKeyAutoGenerated(),
  title: text("title").notNull(),
  img: text("img").notNull(),
});

export const categoryRelations = relations(category, ({ many }) => ({
  products: many(products),
}));

export const users = sqliteTable("user", {
  id: primaryKeyAutoGenerated(),
  email: text("email").notNull().unique(),
  displayName: text("displayName").notNull(),
  createdAt: text("createdAt")
    .notNull()
    .default(sql`CURRENT_TIMESTAMP`),
});